+++ wasm/src/org.graalvm.wasm.test/src/org/graalvm/wasm/test/suites/validation/ValidationSuite.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -314,8 +314,8 @@ public class ValidationSuite extends WasmFileSuite {
                                         "(start 0) (func (result i32) i32.const 42)",
                                         Failure.Type.INVALID),
                         stringCase(
-                                        "Start function - takes arguments",
-                                        "Start function cannot take arguments.",
+                                        "Start function - takes parameters",
+                                        "Start function cannot take parameters.",
                                         "(start 0) (func (param i32))",
                                         Failure.Type.INVALID),