+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/WasmCodeEntry.java
@@ -51,15 +51,19 @@ public final class WasmCodeEntry {
     private final int maxStackSize;
     private final BranchProfile errorBranch = BranchProfile.create();
     @CompilationFinal(dimensions = 1) private final int[] extraData;
+    @CompilationFinal(dimensions = 1) private final byte[] resultTyps;
     private final int numLocals;
+    private final int resultCount;
 
-    public WasmCodeEntry(WasmFunction function, byte[] data, byte[] localTypes, int maxStackSize, int[] extraData) {
+    public WasmCodeEntry(WasmFunction function, byte[] data, byte[] localTypes, byte[] resultTypes, int maxStackSize, int[] extraData) {
         this.function = function;
         this.data = data;
         this.localTypes = localTypes;
         this.maxStackSize = maxStackSize;
         this.extraData = extraData;
         this.numLocals = localTypes.length;
+        this.resultTyps = resultTypes;
+        this.resultCount = resultTypes.length;
     }
 
     public WasmFunction function() {
@@ -88,6 +92,14 @@ public final class WasmCodeEntry {
 
     public int[] extraData() {
         return extraData;
+    }
+
+    public int resultCount() {
+        return resultCount;
+    }
+
+    public byte resultType(int index) {
+        return resultTyps[index];
     }
 
     public void errorBranch() {