+++ wasm/src/org.graalvm.wasm/src/org/graalvm/wasm/WasmMultiValueResult.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -39,45 +39,11 @@
  * SOFTWARE.
  */
 
-package org.graalvm.wasm.api;
+package org.graalvm.wasm;
 
-import org.graalvm.wasm.ModuleLimits;
+public final class WasmMultiValueResult {
+    public static final WasmMultiValueResult INSTANCE = new WasmMultiValueResult();
 
-public final class JsConstants {
-    private JsConstants() {
+    private WasmMultiValueResult() {
     }
-
-    // Limits specified by https://www.w3.org/TR/wasm-js-api/#limits
-    private static final int MODULE_SIZE_LIMIT = 1 << 30;
-    private static final int TYPE_COUNT_LIMIT = 1000000;
-    private static final int FUNCTION_COUNT_LIMIT = 1000000;
-    private static final int IMPORT_COUNT_LIMIT = 100000;
-    private static final int EXPORT_COUNT_LIMIT = 100000;
-    private static final int GLOBAL_COUNT_LIMIT = 1000000;
-    private static final int DATA_SEGMENT_LIMIT = 100000;
-    private static final int ELEMENT_SEGMENT_LIMIT = 10000000;
-    private static final int FUNCTION_SIZE_LIMIT = 7654321;
-    private static final int PARAM_COUNT_LIMIT = 1000;
-    private static final int RESULT_COUNT_LIMIT = 1;
-    private static final int MULTI_VALUE_RESULT_COUNT_LIMIT = 1000;
-    private static final int LOCAL_COUNT_LIMIT = 50000;
-    private static final int TABLE_SIZE_LIMIT = 10000000;
-    private static final int MEMORY_SIZE_LIMIT = 32767;
-
-    static final ModuleLimits JS_LIMITS = new ModuleLimits(
-                    MODULE_SIZE_LIMIT,
-                    TYPE_COUNT_LIMIT,
-                    FUNCTION_COUNT_LIMIT,
-                    IMPORT_COUNT_LIMIT,
-                    EXPORT_COUNT_LIMIT,
-                    GLOBAL_COUNT_LIMIT,
-                    DATA_SEGMENT_LIMIT,
-                    ELEMENT_SEGMENT_LIMIT,
-                    FUNCTION_SIZE_LIMIT,
-                    PARAM_COUNT_LIMIT,
-                    RESULT_COUNT_LIMIT,
-                    MULTI_VALUE_RESULT_COUNT_LIMIT,
-                    LOCAL_COUNT_LIMIT,
-                    TABLE_SIZE_LIMIT,
-                    MEMORY_SIZE_LIMIT);
 }